Split controlmodel_ipadapter file
